move all interfaces to types.ts
