+++ src/types.ts
@@ -25,3 +25,41 @@ export interface ILeafHeatDimensions {
   heat_outerRadius: number;
   node_outerRadius: number;
 }
+
+export interface CarouselProps {
+  isPlaying: boolean;
+  isShown: boolean;
+  togglePlaying: () => void;
+  toggleShown: () => void;
+}
+
+export interface SwatchesProps {
+  count: number;
+  colorScale: d3.ScaleSequential<string, never>;
+}
+
+export interface TreeNodeProps {
+  treeNode: d3.HierarchyNode<ITreeNode>;
+}
+
+export interface NodeLinkProps {
+  linkData: d3.HierarchyNode<ITreeNode>;
+}
+
+export interface ITreeNode {
+  id: string;
+  x: number;
+  y: number;
+  angle: number;
+  children?: ITreeNode[];
+}
+
+export interface TextAngleParams {
+  startAngle: number;
+  endAngle: number;
+  distance: string;
+  depth: number;
+  id: string;
+  hovered: boolean;
+  selected: boolean;
+}
